testing 16 bit float

+++ freesound.py
@@ -10,11 +10,11 @@ import shutil
 import pandas as pd
 import librosa
 import numpy as np
-import keras
+from tensorflow import keras
 import GPyOpt
 from sklearn.cross_validation import StratifiedKFold
-from keras import losses, models, optimizers
-from keras.activations import relu, softmax
+from tensorflow.keras import losses, models, optimizers
+from tensorflow.keras.activations import relu, softmax
 import tensorflow as tf
 
 np.random.seed(1001)
@@ -66,22 +66,22 @@ class SoundClassifier(object):
         else:
             nclass = 41
             inp = keras.layers.Input(tensor=input_tensor)
-        x = keras.layers.Convolution2D(32, (4, 10), padding="same", dtype=tf.float16, input_dtype=tf.float16)(inp)
+        x = keras.layers.Convolution2D(32, (4, 10), padding="same")(inp)
         x = keras.layers.BatchNormalization()(x)
         x = keras.layers.Activation("relu")(x)
         x = keras.layers.MaxPool2D()(x)
 
-        x = keras.layers.Convolution2D(32, (4, 10), padding="same", dtype=tf.float16, input_dtype=tf.float16)(x)
+        x = keras.layers.Convolution2D(32, (4, 10), padding="same")(x)
         x = keras.layers.BatchNormalization()(x)
         x = keras.layers.Activation("relu")(x)
         x = keras.layers.MaxPool2D()(x)
 
-        x = keras.layers.Convolution2D(32, (4, 10), padding="same", dtype=tf.float16, input_dtype=tf.float16)(x)
+        x = keras.layers.Convolution2D(32, (4, 10), padding="same")(x)
         x = keras.layers.BatchNormalization()(x)
         x = keras.layers.Activation("relu")(x)
         x = keras.layers.MaxPool2D()(x)
 
-        x = keras.layers.Convolution2D(32, (4, 10), padding="same", dtype=tf.float16, input_dtype=tf.float16)(x)
+        x = keras.layers.Convolution2D(32, (4, 10), padding="same")(x)
         x = keras.layers.BatchNormalization()(x)
         x = keras.layers.Activation("relu")(x)
         x = keras.layers.MaxPool2D()(x)
